requirements.txt
requests>=2.25.1
unidecode>=1.3.2
goose3>=3.1.9
torch 
torchvision 
transformers 
datasets 
scikit-learn 
Pillow 
PyYAML